requirements.txt
python-telegram-bot
beautifulsoup4
requests
apscheduler
python-dotenv